third_party/WebKit/LayoutTests/fast/dom/inline-event-attributes-event-param-name-expected.txt

Tests the name of the event parameter for inline event attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS type is "object"
PASS type is "undefined"
PASS type is "undefined"
PASS type is "object"
PASS successfullyParsed is true

TEST COMPLETE
